Deletion: WeeklySchedule moved to scheduler.models
