+++ src/main/java/com/iwm/backend/schedulegenerator/FuzzyGeneticScheduleGenerator.java
@@ -3,7 +3,7 @@ package com.iwm.backend.schedulegenerator;
 import com.iwm.backend.schedulegenerator.configurations.FGAConfigs;
 import com.iwm.backend.schedulegenerator.models.Employee;
 import com.iwm.backend.schedulegenerator.models.Population;
-import com.iwm.backend.api.models.WeeklySchedule;
+import com.iwm.backend.schedulegenerator.models.WeeklySchedule;
 import com.iwm.backend.schedulegenerator.models.Shift;
 import com.iwm.backend.schedulegenerator.util.CalculationsUtility;
 import com.iwm.backend.trial.EmloyeesReader;
