+++ src/main/java/com/iwm/backend/schedulegenerator/Test.java
@@ -1,7 +1,7 @@
 package com.iwm.backend.schedulegenerator;
 
 import com.iwm.backend.schedulegenerator.models.Shift;
-import com.iwm.backend.api.models.WeeklySchedule;
+import com.iwm.backend.schedulegenerator.models.WeeklySchedule;
 import com.iwm.backend.schedulegenerator.models.*;
 
 import java.time.LocalDate;
