+++ src/main/java/com/iwm/backend/schedulegenerator/models/Population.java
@@ -1,6 +1,5 @@
 package com.iwm.backend.schedulegenerator.models;
 
-import com.iwm.backend.api.models.WeeklySchedule;
 import com.iwm.backend.schedulegenerator.configurations.FGAConfigs;
 import com.iwm.backend.trial.DemandReader;
 import com.iwm.backend.trial.EmloyeesReader;
